+++ lib/common/widgets/chips/animated_switch.dart
@@ -51,10 +51,11 @@ class TAnimatedToggleSwitch extends StatelessWidget {
     this.indicatorSize = const Size.fromWidth(25),
     this.boxShadow = const [
       BoxShadow(
-          color: Colors.black26,
-          spreadRadius: 1,
-          blurRadius: 2,
-          offset: Offset(0, 1.5))
+        color: Colors.black26,
+        spreadRadius: 1,
+        blurRadius: 2,
+        offset: Offset(0, 1.5),
+      )
     ],
   });
 
@@ -117,10 +118,7 @@ class TAnimatedToggleSwitch extends StatelessWidget {
       indicatorSize: indicatorSize,
 
       // Default toggle switch style
-      style: ToggleStyle(
-          backgroundColor: backgroundColor,
-          borderColor: borderColor,
-          boxShadow: boxShadow),
+      style: ToggleStyle(backgroundColor: backgroundColor, borderColor: borderColor, boxShadow: boxShadow),
 
       // Dynamic text builder
       textBuilder: textBuilder ??
@@ -132,17 +130,11 @@ class TAnimatedToggleSwitch extends StatelessWidget {
               ),
 
       // Dynamic icon builder
-      iconBuilder: iconBuilder ??
-          (value) => value
-              ? Icon(Icons.check_circle, color: TColors.primary)
-              : Icon(Icons.circle_outlined),
+      iconBuilder: iconBuilder ?? (value) => value ? Icon(Icons.check_circle, color: TColors.primary) : Icon(Icons.circle_outlined),
 
       // Dynamic style builder for indicator color
       styleBuilder: styleBuilder ??
-          (value) => ToggleStyle(
-              indicatorColor: value
-                  ? activeColor ?? TColors.primaryBackground
-                  : inactiveColor ?? TColors.grey),
+          (value) => ToggleStyle(indicatorColor: value ? activeColor ?? TColors.primaryBackground : inactiveColor ?? TColors.grey),
     );
   }
 }