+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.0.7
+- `animation_loader.dart` assert fixed
+- `THeader` background colors updated
+- Tooltip added in `TIcon` Widget
+- Default `Table Header Row` color changed to `lightBackground`
+- `Toggle Icons` selected Radius fixed
+
 ## 2.0.1
 - Quill Text Editor added. `TTextEditor`
 ## 2.0.0